Split main.cpp
