+++ test_mode.h
@@ -1,0 +1,10 @@
+//
+// Created by kirill on 20.03.2020.
+//
+
+#ifndef UNTITLED_TEST_MODE_H
+#define UNTITLED_TEST_MODE_H
+
+constexpr static bool test_mode = false;
+
+#endif //UNTITLED_TEST_MODE_H
